IvyFEM.dll ver0.0.0.62 3次元電磁導波路の伝達問題、共振器の固有値問題

+++ src/IvyFEM/IvyFEM/EMWaveguide2DHPlaneFEM.cs
@@ -218,7 +218,7 @@ namespace IvyFEM
                 int incidentModeId = -1;
                 if (World.GetIncidentPortId(QuantityId) == portId)
                 {
-                    incidentModeId = (int)World.GetIncidentModeId(QuantityId);
+                    incidentModeId = World.GetIncidentModeId(QuantityId);
                 }
                 System.Numerics.Complex[] S1 = eigenFEM.CalcSMatrix(omega, incidentModeId, betas, ezEVecs, portEz);
                 S[portId] = S1;
